+++ command/agent/agent.go
@@ -434,9 +434,9 @@ func (a *Agent) clientConfig() (*clientconfig.Config, error) {
 		res = new(structs.NodeReservedResources)
 		conf.Node.ReservedResources = res
 	}
-	res.Cpu.CpuShares = uint64(a.config.Client.Reserved.CPU)
-	res.Memory.MemoryMB = uint64(a.config.Client.Reserved.MemoryMB)
-	res.Disk.DiskMB = uint64(a.config.Client.Reserved.DiskMB)
+	res.Cpu.CpuShares = a.config.Client.Reserved.CPU
+	res.Memory.MemoryMB = a.config.Client.Reserved.MemoryMB
+	res.Disk.DiskMB = a.config.Client.Reserved.DiskMB
 	res.Networks.ReservedHostPorts = a.config.Client.Reserved.ReservedPorts
 
 	conf.Version = a.config.Version
